# packages/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: RafayPair API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: >-
    Milestone 1 contract for authentication, two-person pairing, directional consent,
    care requests, privacy pause, notification registration, and authenticated realtime delivery.
    Browser clients use secure cookies plus double-submit CSRF. Native clients use opaque bearer tokens.
...
servers:
  - url: https://api.rafaypair.com
    description: Production
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /v1/pairs/current:
    get:
      tags: [Pairs]
      summary: Read the current waiting or active pair
      operationId: getCurrentPair
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Current pair
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PairResponse" }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
    post:
      tags: [Pairs]
      summary: Create a waiting pair and one-time join code
      operationId: createPair
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/CsrfTokenOptionalForNative"
      responses:
        "201":
          description: Waiting pair. joinCode is returned only by this operation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PairResponse" }
        "401": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
    delete:
      tags: [Pairs]
      summary: Disconnect the pair and durably revoke all partner grants
      operationId: disconnectPair
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/CsrfTokenOptionalForNative"
      responses:
        "204": { description: Pair disconnected and partner access revoked }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
  /v1/pairs/join:
    post:
      tags: [Pairs]
      summary: Join a waiting pair as its second and final member
      operationId: joinPair
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/CsrfTokenOptionalForNative"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JoinPairRequest" }
      responses:
        "200":
          description: Active pair; every directional consent starts denied.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PairResponse" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
...
  /v1/integrity/android/challenges:
    post:
      tags: [Integrity]
      summary: Mint a short-lived Android integrity binding challenge
      operationId: createAndroidIntegrityChallenge
      description: >-
        Android only. The client hashes the documented canonical action with this one-time
        challenge before requesting a Standard Integrity token. A verdict is a risk signal
        and never grants or revokes API authorization by itself.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ClientPlatform"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              { $ref: "#/components/schemas/AndroidIntegrityChallengeRequest" }
      responses:
        "201":
          description: Short-lived challenge bound to the authenticated Android session family.
          content:
            application/json:
              schema:
                {
                  $ref: "#/components/schemas/AndroidIntegrityChallengeResponse",
                }
        "400": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "501": { $ref: "#/components/responses/Problem" }
  /v1/integrity/android/assessments:
    post:
      tags: [Integrity]
      summary: Decode and record a content-bound Standard Integrity verdict
      operationId: submitAndroidIntegrityAssessment
      description: >-
        Android only. The encrypted provider token is sent directly to Google Play from the
        backend, never persisted, and checked for package, signing certificate, minimum app
        version, request hash, and freshness before its sanitized risk metadata is recorded.
        Google Standard Integrity automatic replay protection remains enabled. The result is
        not an authorization decision.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ClientPlatform"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              { $ref: "#/components/schemas/AndroidIntegrityAssessmentRequest" }
      responses:
        "202":
          description: Sanitized risk signal recorded; authorization is unchanged.
          content:
            application/json:
              schema:
                {
                  $ref: "#/components/schemas/AndroidIntegrityAssessmentResponse",
                }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "501": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Native-only short-lived opaque access token stored in Keychain or Keystore.
    cookieAuth:
      type: apiKey
      in: cookie
      name: rafay_access
      description: Web-only HttpOnly Secure SameSite=Strict access cookie.
    realtimeTicket:
      type: apiKey
      in: header
      name: Sec-WebSocket-Protocol
      description: >-
        One-time, 30-second ticket bound to the authenticated user, session, and pair. It is offered as
        rafaypair.ticket.<base64url-ticket> beside rafaypair.v1; only rafaypair.v1 is selected.
  parameters:
    ClientPlatform:
      name: X-Rafay-Client
      in: header
      required: true
      schema:
        type: string
        enum: [web, ios, android]
    CsrfTokenOptionalForNative:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required for Web state-changing requests; must equal the readable rafay_csrf cookie.
      schema: { type: string }
  responses:
    Problem:
      description: RFC 9457 / RFC 7807 problem details
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
  schemas:
...
    PairMember:
      type: object
      additionalProperties: false
      required: [userId, displayName, joinedAt]
      properties:
        userId: { type: string, format: uuid }
        displayName: { type: string, minLength: 1, maxLength: 80 }
        joinedAt: { type: string, format: date-time }
    Pair:
      type: object
      additionalProperties: false
      required: [id, status, members, createdAt]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [waiting, active] }
        members:
          type: array
          minItems: 1
          maxItems: 2
          items: { $ref: "#/components/schemas/PairMember" }
        joinCode:
          type: string
          pattern: "^[A-Z2-9]{8}$"
          description: Present only in the create response while the pair is waiting.
        createdAt: { type: string, format: date-time }
    PairResponse:
      type: object
      additionalProperties: false
      required: [pair]
      properties:
        pair: { $ref: "#/components/schemas/Pair" }
    JoinPairRequest:
      type: object
      additionalProperties: false
      required: [code]
      properties:
        code: { type: string, pattern: "^[A-Z2-9]{8}$" }
...
    AndroidIntegrityChallengeRequest:
      type: object
      additionalProperties: false
      required: [action]
      properties:
        action: { type: string, enum: [session_start] }
    AndroidIntegrityChallengeResponse:
      type: object
      additionalProperties: false
      required: [challenge]
      properties:
        challenge:
          type: object
          additionalProperties: false
          required: [id, action, bindingVersion, expiresAt]
          properties:
            id: { type: string, format: uuid }
            action: { type: string, enum: [session_start] }
            bindingVersion: { type: string, const: sha256-v1 }
            expiresAt: { type: string, format: date-time }
    AndroidIntegrityAssessmentRequest:
      type: object
      additionalProperties: false
      required: [challengeId, action, integrityToken]
      properties:
        challengeId: { type: string, format: uuid }
        action: { type: string, enum: [session_start] }
        integrityToken:
          type: string
          minLength: 64
          maxLength: 60000
          pattern: '^\S+$'
          description: Opaque encrypted Standard Integrity token; never stored or logged.
    AndroidIntegrityAssessmentResponse:
      type: object
      additionalProperties: false
      required: [assessment]
      properties:
        assessment:
          type: object
          additionalProperties: false
          required: [id, signal, evaluatedAt]
          properties:
            id: { type: string, format: uuid }
            signal:
              type: string
              enum: [low_risk, elevated_risk, invalid_binding]
              description: Risk telemetry only; it does not alter authorization.
            evaluatedAt: { type: string, format: date-time }
...
    ProblemDetails:
      type: object
      additionalProperties: true
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string }
        code: { type: string }
        requestId: { type: string }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }